Added persistence and registered classes with Ninject
Added persistence through entity framework, and set up Ninject to bind
an interface to a class. This involved:
-Creating a repository interface (ICustomerRepository) for saving
customers, and a concrete repository for using the interface
(EFCustomerRepository).
-Created a context for EF to use (WEBAPPS)
-Adding a Ninject Dependency Resolver which maps the customer repository
interface to the EF implementation, and setting up the app to use that
resolver through NinjectWebCommon.
-Updating the Customer class to have a CustomerID (it didn't before),
and a table name for when EF does its work.
-Finally, putting this to work in the app in the Home Controller.